11731 kaio: cast between incompatible function types Reviewed by: Andy Fiddaman <omnios@citrus-it.co.uk> Approved by: Dan McDonald <danmcd@joyent.com>

+++ usr/src/uts/common/os/aio.c
@@ -143,7 +143,7 @@ static int aio_port_callback(void *, int *, pid_t, int, void *);
 static struct sysent kaio_sysent = {
 	6,
 	SE_NOUNLOAD | SE_64RVAL | SE_ARGC,
-	(int (*)())kaioc
+	(int (*)())(uintptr_t)kaioc
 };
 
 #ifdef _SYSCALL32_IMPL
